requirements.txt
python-dotenv
requests
smtplib
email-validator